auth-service/requirements.txt
fastapi[standard]
pydantic
passlib[argon2]
pyjwt
python-dotenv